+++ hyperdbg/hprdbgctrl/code/debugger/core/debugger.cpp
@@ -538,7 +538,7 @@ DebuggerGetNtoskrnlBase()
 
     NtQuerySystemInformation(SystemModuleInformation, Modules, SysModuleInfoBufferSize, NULL);
 
-    for (int i = 0; i < Modules->NumberOfModules; i++)
+    for (UINT32 i = 0; i < Modules->NumberOfModules; i++)
     {
         if (!strcmp((const char *)Modules->Modules[i].FullPathName + Modules->Modules[i].OffsetToFileName,
                     "ntoskrnl.exe"))
@@ -720,11 +720,11 @@ InterpretScript(vector<string> * SplittedCommand,
     string         AppendedFinalBuffer;
     vector<string> SaveBuffer;
     vector<int>    IndexesToRemove;
-    int            Index            = 0;
-    int            NewIndexToRemove = 0;
-    int            OpenBracket      = 0;
-    size_t         CountOfOpenBrackets;
-    size_t         CountOfCloseBrackets;
+    UINT32         Index            = 0;
+    UINT32         NewIndexToRemove = 0;
+    UINT32         OpenBracket      = 0;
+    UINT32         CountOfOpenBrackets;
+    UINT32         CountOfCloseBrackets;
     UINT32         IndexInCommandCaseSensitive          = 0;
     vector<string> SplittedCommandCaseSensitiveInstance = *SplittedCommandCaseSensitive;
     string         TempStr;
@@ -794,7 +794,7 @@ InterpretScript(vector<string> * SplittedCommand,
                 //
                 // Check the count of brackets in the string and add it to OpenBracket
                 //
-                size_t CountOfBrackets = count(Section.begin(), Section.end(), '{');
+                UINT32 CountOfBrackets = count(Section.begin(), Section.end(), '{');
 
                 //
                 // Add it to the open brackets
@@ -814,7 +814,7 @@ InterpretScript(vector<string> * SplittedCommand,
                 //
                 // Check the count of brackets in the string and add it to OpenBracket
                 //
-                size_t CountOfBrackets = count(Section.begin(), Section.end(), '}');
+                UINT32 CountOfBrackets = count(Section.begin(), Section.end(), '}');
 
                 //
                 // Add it to the open brackets
@@ -871,7 +871,7 @@ InterpretScript(vector<string> * SplittedCommand,
                 //
                 // Check the count of brackets in the string and add it to OpenBracket
                 //
-                size_t CountOfBrackets = count(Section.begin(), Section.end(), '{');
+                UINT32 CountOfBrackets = count(Section.begin(), Section.end(), '{');
 
                 //
                 // Add it to the open brackets (-1 because script starts with { which
@@ -909,7 +909,7 @@ InterpretScript(vector<string> * SplittedCommand,
                 //
                 // Check the count of brackets in the string and add it to OpenBracket
                 //
-                size_t CountOfBrackets = count(Section.begin(), Section.end(), '}');
+                UINT32 CountOfBrackets = count(Section.begin(), Section.end(), '}');
 
                 //
                 // Add it to the open brackets
@@ -1029,7 +1029,7 @@ InterpretScript(vector<string> * SplittedCommand,
                     //
                     // Check the count of brackets in the string and add it to OpenBracket
                     //
-                    size_t CountOfBrackets = count(Section.begin(), Section.end(), '}');
+                    UINT32 CountOfBrackets = count(Section.begin(), Section.end(), '}');
 
                     //
                     // Add it to the open brackets
@@ -1530,7 +1530,7 @@ InterpretConditionsAndCodes(vector<string> * SplittedCommand,
     // Set the buffer and length
     //
     *BufferAddress = (UINT64)FinalBuffer;
-    *BufferLength  = ParsedBytes.size();
+    *BufferLength  = (UINT32)ParsedBytes.size();
 
     //
     // Removing the code or condition indexes from the command
@@ -1909,7 +1909,7 @@ SendEventToKernel(PDEBUGGER_GENERAL_EVENT_DETAIL Event,
             //
             if (g_IsConnectedToRemoteDebuggee)
             {
-                RemoteConnectionSendCommand("g", strlen("g") + 1);
+                RemoteConnectionSendCommand("g", (UINT32)strlen("g") + 1);
             }
 
             ShowMessages("\n");